requirements.txt
boto3>=1.28.57
awscli>=1.29.57
botocore>=1.31.57
qdrant-client
python-dotenv
requests-aws4auth
